requirements.txt
plyfile
tqdm
timm 
opencv-python
torchmetrics
trimesh
h5py
scipy
kornia
ninja
matplotlib
datasets
diffusers==0.10.2
transformers==4.26.0
accelerate
toml
albumentations
einops
pytorch_lightning==2.3
wget